# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CUI API",
    "version": "0.1.0"
  },
...
  "paths": {
...
    "/api/v1/runs/{runId}/cancellation": {
      "post": {
        "summary": "Cancel a run",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "runId",
            "in": "path"
          }
        ],
        "responses": {
          "202": {
            "description": "The running run was cancelled.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": ["ok"]
                    }
                  },
                  "required": ["status"]
                }
              }
            }
          },
          "404": {
            "description": "Error response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "409": {
            "description": "Error response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{runId}/events": {
      "get": {
        "summary": "Stream run events",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "runId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Server-sent event stream. Each event payload matches RunStreamEvent.",
            "content": {
              "text/event-stream": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "enum": ["run.output.delta"]
                        },
                        "text": {
                          "type": "string"
                        }
                      },
                      "required": ["type", "text"]
                    },
                    {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "enum": ["run.trace"]
                        },
                        "event": {}
                      },
                      "required": ["type"]
                    },
                    {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "enum": ["session.updated"]
                        },
                        "session": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "origin": {
                              "type": "string",
                              "enum": ["chat", "shell"]
                            },
                            "workspace": {
                              "type": "string"
                            },
                            "title": {
                              "type": "string"
                            },
                            "summary": {
                              "type": "string"
                            },
                            "doneAt": {
                              "type": "string",
                              "format": "date-time"
                            },
                            "createdAt": {
                              "type": "string",
                              "format": "date-time"
                            },
                            "updatedAt": {
                              "type": "string",
                              "format": "date-time"
                            },
                            "messages": {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "id": {
                                    "type": "string"
                                  },
                                  "role": {
                                    "type": "string",
                                    "enum": ["assistant", "user"]
                                  },
                                  "kind": {
                                    "type": "string",
                                    "enum": ["response", "trace"]
                                  },
                                  "round": {
                                    "type": "integer",
                                    "exclusiveMinimum": 0
                                  },
                                  "content": {
                                    "type": "string"
                                  },
                                  "createdAt": {
                                    "type": "string",
                                    "format": "date-time"
                                  }
                                },
                                "required": ["id", "role", "content", "createdAt"]
                              }
                            },
                            "rounds": {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "round": {
                                    "type": "integer",
                                    "exclusiveMinimum": 0
                                  },
                                  "hasChanges": {
                                    "type": "boolean"
                                  },
                                  "createdAt": {
                                    "type": "string",
                                    "format": "date-time"
                                  },
                                  "atomicReviewStatus": {
                                    "anyOf": [
                                      {
                                        "type": "string",
                                        "enum": ["ready"]
                                      },
                                      {
                                        "type": "string",
                                        "enum": ["failed"]
                                      }
                                    ]
                                  }
                                },
                                "required": ["round", "hasChanges", "createdAt"]
                              }
                            },
                            "queuedPrompts": {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "id": {
                                    "type": "string"
                                  },
                                  "mode": {
                                    "type": "string",
                                    "enum": ["chat", "shell"]
                                  },
                                  "prompt": {
                                    "type": "string"
                                  },
                                  "createdAt": {
                                    "type": "string",
                                    "format": "date-time"
                                  }
                                },
                                "required": ["id", "mode", "prompt", "createdAt"]
                              }
                            },
                            "currentRound": {
                              "type": "integer",
                              "minimum": 0
                            },
                            "gitBranch": {
                              "type": "string"
                            },
                            "isRunning": {
                              "type": "boolean"
                            },
                            "runningRunId": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "id",
                            "workspace",
                            "title",
                            "createdAt",
                            "updatedAt",
                            "messages",
                            "currentRound",
                            "isRunning"
                          ]
                        }
                      },
                      "required": ["type", "session"]
                    },
                    {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "enum": ["run.succeeded"]
                        },
                        "session": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "origin": {
                              "type": "string",
                              "enum": ["chat", "shell"]
                            },
                            "workspace": {
                              "type": "string"
                            },
                            "title": {
                              "type": "string"
                            },
                            "summary": {
                              "type": "string"
                            },
                            "doneAt": {
                              "type": "string",
                              "format": "date-time"
                            },
                            "createdAt": {
                              "type": "string",
                              "format": "date-time"
                            },
                            "updatedAt": {
                              "type": "string",
                              "format": "date-time"
                            },
                            "messages": {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "id": {
                                    "type": "string"
                                  },
                                  "role": {
                                    "type": "string",
                                    "enum": ["assistant", "user"]
                                  },
                                  "kind": {
                                    "type": "string",
                                    "enum": ["response", "trace"]
                                  },
                                  "round": {
                                    "type": "integer",
                                    "exclusiveMinimum": 0
                                  },
                                  "content": {
                                    "type": "string"
                                  },
                                  "createdAt": {
                                    "type": "string",
                                    "format": "date-time"
                                  }
                                },
                                "required": ["id", "role", "content", "createdAt"]
                              }
                            },
                            "rounds": {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "round": {
                                    "type": "integer",
                                    "exclusiveMinimum": 0
                                  },
                                  "hasChanges": {
                                    "type": "boolean"
                                  },
                                  "createdAt": {
                                    "type": "string",
                                    "format": "date-time"
                                  },
                                  "atomicReviewStatus": {
                                    "anyOf": [
                                      {
                                        "type": "string",
                                        "enum": ["ready"]
                                      },
                                      {
                                        "type": "string",
                                        "enum": ["failed"]
                                      }
                                    ]
                                  }
                                },
                                "required": ["round", "hasChanges", "createdAt"]
                              }
                            },
                            "queuedPrompts": {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "id": {
                                    "type": "string"
                                  },
                                  "mode": {
                                    "type": "string",
                                    "enum": ["chat", "shell"]
                                  },
                                  "prompt": {
                                    "type": "string"
                                  },
                                  "createdAt": {
                                    "type": "string",
                                    "format": "date-time"
                                  }
                                },
                                "required": ["id", "mode", "prompt", "createdAt"]
                              }
                            },
                            "currentRound": {
                              "type": "integer",
                              "minimum": 0
                            },
                            "gitBranch": {
                              "type": "string"
                            },
                            "isRunning": {
                              "type": "boolean"
                            },
                            "runningRunId": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "id",
                            "workspace",
                            "title",
                            "createdAt",
                            "updatedAt",
                            "messages",
                            "currentRound",
                            "isRunning"
                          ]
                        }
                      },
                      "required": ["type", "session"]
                    },
                    {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "enum": ["run.failed"]
                        },
                        "error": {
                          "type": "string"
                        }
                      },
                      "required": ["type", "error"]
                    },
                    {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "enum": ["run.cancelled"]
                        }
                      },
                      "required": ["type"]
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Error response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        }
      }
    }
  },
...
}